[TESTS] Extract common methods for socket mocks to reusable base classes
