+++ tests/unit_test/test_async/test_async_api/test_client/test_tcp.py
@@ -64,11 +64,10 @@ class TestAsyncTCPNetworkClient(BaseTestClient):
         socket_family: int,
         global_local_address: tuple[str, int],
     ) -> tuple[str, int]:
-        cls.set_local_address_to_socket_mock(
+        cls.set_local_address_to_async_socket_adapter_mock(
             mock_stream_socket_adapter,
             socket_family,
             global_local_address,
-            "get_local_address",
         )
         return global_local_address
 
@@ -80,11 +79,10 @@ class TestAsyncTCPNetworkClient(BaseTestClient):
         socket_family: int,
         global_remote_address: tuple[str, int],
     ) -> tuple[str, int]:
-        cls.set_remote_address_to_socket_mock(
+        cls.set_remote_address_to_async_socket_adapter_mock(
             mock_stream_socket_adapter,
             socket_family,
             global_remote_address,
-            "get_remote_address",
         )
         return global_remote_address
